requirements.txt
numpy
pytorch
fire
scipy
gym
